+++ crates/pallets/starknet/src/blockifier_state_adapter.rs
@@ -2,13 +2,12 @@ use alloc::collections::{BTreeMap, BTreeSet};
 use core::marker::PhantomData;
 
 use blockifier::execution::contract_class::ContractClass;
-use blockifier::state::cached_state::{CommitmentStateDiff, ContractStorageKey};
+use blockifier::state::cached_state::{CommitmentStateDiff, ContractStorageKey, StateChangesCount};
 use blockifier::state::errors::StateError;
 use blockifier::state::state_api::{State, StateReader, StateResult};
 use indexmap::IndexMap;
 use mp_felt::Felt252Wrapper;
-use mp_state::{FeeConfig, StateChanges};
-use sp_core::Get;
+use mp_state::StateChanges;
 use starknet_api::api_core::{ClassHash, CompiledClassHash, ContractAddress, Nonce};
 use starknet_api::hash::StarkFelt;
 use starknet_api::state::StorageKey;
@@ -32,19 +31,15 @@ impl<T> StateChanges for BlockifierStateAdapter<T>
 where
     T: Config,
 {
-    fn count_state_changes(&self) -> (usize, usize, usize, usize) {
+    fn count_state_changes(&self) -> StateChangesCount {
         let keys = self.storage_update.keys();
         let n_contract_updated = BTreeSet::from_iter(keys.clone().map(|&(contract_address, _)| contract_address)).len();
-        (n_contract_updated, keys.len(), self.class_hash_update, self.compiled_class_hash_update)
-    }
-}
-
-impl<T> FeeConfig for BlockifierStateAdapter<T>
-where
-    T: Config,
-{
-    fn is_transaction_fee_disabled(&self) -> bool {
-        T::DisableTransactionFee::get()
+        StateChangesCount {
+            n_modified_contracts: n_contract_updated,
+            n_storage_updates: keys.len(),
+            n_class_hash_updates: self.class_hash_update,
+            n_compiled_class_hash_updates: self.compiled_class_hash_update,
+        }
     }
 }