+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 - chore(data-availability-avail): implement fire and forget, and add ws
   reconnection logic
 - chore: update `polkadot-sdk` to `release-polkadot-v1.3.0`
+- refactor: use `map` in `estimate_fee` to stop computation on error
 
 ## v0.5.0