+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next release
 
+- fix: Change seliazation of bitvec to &[u8] in merkle tree to avoid memory
+  uninitialized
 - chore: change SCARB config version for foundry CI
 
 ## v0.6.0