+++ CHANGELOG.md
@@ -2,13 +2,13 @@
 
 ## Next release
 
+- feat(rpc) implemented v0.5.1 missing fields + bumped starknet.rs to match them
 - fix: Change seliazation of bitvec to &[u8] in merkle tree to avoid memory
   uninitialized
 - chore: change SCARB config version for foundry CI
 
 ## v0.6.0
 
-- feat(rpc) implemented v0.5.1 missing fields + bumped starknet.rs to match them
 - chore: release v0.6.0
 - refacto: substrate/starknet names in rpc library
 - feat(rpc): Added starknet_getTransactionStatus and removed