+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Next release
 
-- fix: Change seliazation of bitvec to &[u8] in merkle tree to avoid memory
+- feat(rpc): Added starknet_simulateTransactions
+- fix: Change serialization of bitvec to &[u8] in merkle tree to avoid memory
   uninitialized
 - chore: change SCARB config version for foundry CI